requirements.txt
black

numpy
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
pillow
PyGLM
